Refactor hardhat-ethers.
Add a types module that defines the interface for additional Hardhat
fields.
Separate out compilation projects for the plugin and its tests.
