+++ packages/hardhat-ethers/test/fixture-projects/hardhat-project/hardhat.config.js
@@ -1,4 +1,4 @@
-require("../../../src/index");
+require("../../../src/plugin/index");
 
 module.exports = {
   solidity: "0.5.15",
