+++ packages/hardhat-ethers/test/helpers.ts
@@ -2,6 +2,9 @@ import { resetHardhatContext } from "hardhat/plugins-testing";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 import path from "path";
 
+// Import this plugin type extensions for the HardhatRuntimeEnvironment
+import type {} from "../src/plugin/index";
+
 declare module "mocha" {
   interface Context {
     env: HardhatRuntimeEnvironment;
